Hacked in a 'nearest' method of resampling using scipy's griddata. Customized griddata to take a distance_upper_bound keyword. Also added removable patterns to DRRTV frontend.

+++ py/polar2grid/polar2grid/remap.py
@@ -218,15 +218,164 @@ def run_fornav_c(*args, **kwargs):
                 )
     return result
 
-def run_fornav_py():
-    pass
+import numpy
+# from scipy.interpolate import griddata
+from scipy.interpolate.interpnd import LinearNDInterpolator, NDInterpolatorBase, CloughTocher2DInterpolator, _ndim_coords_from_arrays
+from scipy.spatial import cKDTree
+from polar2grid.core import Workspace
+
+class NearestNDInterpolator(NDInterpolatorBase):
+    def __init__(self, x, y):
+        x = _ndim_coords_from_arrays(x)
+        self._check_init_shape(x, y)
+        self.tree = cKDTree(x)
+        self.points = x
+        self.values = y
+
+    def __call__(self, *args, **kwargs):
+        """
+        Evaluate interpolator at given points.
+
+        Parameters
+        ----------
+        xi : ndarray of float, shape (..., ndim)
+            Points where to interpolate data at.
+
+        """
+        # DJH: Added kwargs and passed them to self.tree.query
+        xi = _ndim_coords_from_arrays(args)
+        xi = self._check_call_shape(xi)
+        dist, i = self.tree.query(xi, **kwargs)
+        max_idx = max(self.values.shape)
+        i[i >= max_idx] = max_idx - 1
+        return self.values[i]
+
+def griddata(points, values, xi, method='linear', fill_value=numpy.nan, distance_upper_bound=2.0):
+    # DJH: Added distance_upper_bound keyword
+    # FIXME: The distance_upper_bound keyword should default to what the KDTree defaults to (inf) and this functionality should probably be put in scipy itself
+    points = _ndim_coords_from_arrays(points)
+
+    if points.ndim < 2:
+        ndim = points.ndim
+    else:
+        ndim = points.shape[-1]
+
+    if ndim == 1 and method in ('nearest', 'linear', 'cubic'):
+        from scipy.interpolate import interp1d
+        points = points.ravel()
+        if isinstance(xi, tuple):
+            if len(xi) != 1:
+                raise ValueError("invalid number of dimensions in xi")
+            xi, = xi
+        # Sort points/values together, necessary as input for interp1d
+        idx = numpy.argsort(points)
+        points = points[idx]
+        values = values[idx]
+        ip = interp1d(points, values, kind=method, axis=0, bounds_error=False,
+                      fill_value=fill_value)
+        return ip(xi)
+    elif method == 'nearest':
+        ip = NearestNDInterpolator(points, values)
+        return ip(xi, distance_upper_bound=distance_upper_bound)
+    elif method == 'linear':
+        ip = LinearNDInterpolator(points, values, fill_value=fill_value)
+        return ip(xi)
+    elif method == 'cubic' and ndim == 2:
+        ip = CloughTocher2DInterpolator(points, values, fill_value=fill_value)
+        return ip(xi)
+    else:
+        raise ValueError("Unknown interpolation method %r for "
+                         "%d dimensional data" % (method, ndim))
+
+def _add_array_border(arr, border_value=-999.0):
+    orig_rows,orig_cols = arr.shape
+    empty_arr = numpy.empty((orig_rows+2, orig_cols+2), dtype=arr.dtype)
+    empty_arr[1:orig_rows+1, 1:orig_cols+1] = arr[:]
+    empty_arr[0, :] = border_value
+    empty_arr[-1, :] = border_value
+    empty_arr[:, 0] = border_value
+    empty_arr[:, -1] = border_value
+    return empty_arr
+
+def _add_array_border2(arr, border_value=-999.0):
+    orig_rows,orig_cols = arr.shape
+    empty_arr = numpy.empty((orig_rows+2, orig_cols+2), dtype=arr.dtype)
+    empty_arr[1:orig_rows+1, 1:orig_cols+1] = arr[:]
+
+    empty_arr[0, 0] = arr[0, 0]
+    empty_arr[0, 1:-1] = arr[0, :]
+    empty_arr[0, -1] = arr[0, -1]
+
+    empty_arr[-1, 0] = arr[-1, 0]
+    empty_arr[-1, 1:-1] = arr[-1, :]
+    empty_arr[-1, -1] = arr[-1, -1]
+
+    empty_arr[1:-1, 0] = arr[:, 0]
+
+    empty_arr[1:-1, -1] = arr[:, -1]
+
+    return empty_arr
+
+def _run_griddata(
+        num_channels,
+        swath_cols,
+        swath_scans,
+        swaths_rows_per_scan,
+        cols_filename,
+        rows_filename,
+        image_filenames,
+        grid_width,
+        grid_height,
+        output_filenames,
+        **kwargs):
+    method = "nearest"
+    # method = "linear"
+    grid_x, grid_y = numpy.mgrid[:grid_height,:grid_width]
+    W = Workspace('.')
+    cols_array = getattr(W, cols_filename.split('.')[0])
+    rows_array = getattr(W, rows_filename.split('.')[0])
+    # cols_array = _add_array_border2(cols_array, border_value=-999.0)
+    # rows_array = _add_array_border2(rows_array, border_value=-999.0)
+
+    for i_fn,o_fn in zip(image_filenames,output_filenames):
+        image_array = getattr(W, i_fn.split('.')[0])
+        # image_array = _add_array_border(image_array, border_value=-999.0)
+
+        #image_array = getattr(W, i_fn.split('.')[0]).copy()
+        #image_array[image_array == -999.0] = numpy.inf
+        image_mask = (image_array != -999.0) | (cols_array == -999.0) | (rows_array == -999.0)
+        # output_array = griddata((cols_array[image_mask].flatten(),rows_array[image_mask].flatten()), image_array[image_mask].flatten(), (grid_y, grid_x), method=method, fill_value=-999.0)
+        output_array = griddata((cols_array.flatten(), rows_array.flatten()), image_array.flatten(), (grid_y, grid_x), method=method, fill_value=-999.0)
+        output_array.tofile(o_fn)
+
+    return {}
+
+def run_fornav_py(*args, **kwargs):
+    proc_pool = kwargs.pop("pool", None)
+    if proc_pool is None:
+        result = _run_griddata(*args, **kwargs)
+    else:
+        result = proc_pool.apply_async(_run_griddata,
+                args=args,
+                kwds=kwargs
+                )
+    return result
 
 def run_fornav(sat, instrument, nav_set_uid, grid_jobs, ll2cr_output,
         num_procs=1, verbose=False, fornav_d=None, fornav_D=None,
-        fill_value=None, do_single_sample=False):
+        fill_value=None, do_single_sample=False, method='ewa'):
     """Run one of the fornavs and return a dictionary mapping grid_name
     to the fornav remapped image data, among other information.
     """
+    methods = {
+        'ewa' : run_fornav_c,
+        'nearest' : run_fornav_py,
+    }
+    if method not in methods:
+        log.error("Unknown remapping method: '%s'", method)
+        raise ValueError("Unknown remapping method: '%s'" % method)
+    remap_method = methods[method]
+
     # Copy the grid_jobs dict (shallow copy)
     fornav_output = grid_jobs
     fill_value = fill_value or DEFAULT_FILL_VALUE
@@ -263,7 +412,7 @@ def run_fornav(sat, instrument, nav_set_uid, grid_jobs, ll2cr_output,
                 band_info["fbf_remapped"] = output_name
 
             for remap_data_as,fornav_job in fornav_group.items():
-                fornav_job["result"] = run_fornav_c(
+                fornav_job["result"] = remap_method(
                             len(fornav_job["inputs"]),
                             ll2cr_output[grid_name]["swath_cols"],
                             ll2cr_output[grid_name]["scans_out"],
@@ -288,21 +437,26 @@ def run_fornav(sat, instrument, nav_set_uid, grid_jobs, ll2cr_output,
                             pool=proc_pool
                             )
 
-        proc_pool.close()
-        proc_pool.join()
+        if proc_pool is not None:
+            proc_pool.close()
+            proc_pool.join()
     except KeyboardInterrupt:
         # Catch keyboard interrupt during pool processing, see comment at
         # top of try block
         log.debug("Keyboard interrupt during fornav call")
-        proc_pool.terminate()
-        proc_pool.join()
+        if proc_pool is not None:
+            proc_pool.terminate()
+            proc_pool.join()
         raise
     
     # Get all the results
     for grid_name,fornav_group in fornav_jobs.items():
         for remap_data_as,fornav_job in fornav_group.items():
             try:
-                fornav_dict = fornav_job["result"].get()
+                if proc_pool is None:
+                    fornav_dict = fornav_job["result"]
+                else:
+                    fornav_dict = fornav_job["result"].get()
                 for (band_kind, band_id),band_dict in fornav_output[grid_name].items():
                     band_dict.update(fornav_dict)
                 log.debug("Fornav successfully completed for grid %s, %s data" % (grid_name,remap_data_as))
@@ -329,7 +483,7 @@ def remap_bands(sat, instrument, nav_set_uid, lon_fbf, lat_fbf,
         grid_jobs, num_procs=1, fornav_d=None, fornav_D=None, forced_gpd=None,
         lat_south=None, lat_north=None, lon_west=None, lon_east=None,
         lat_fill_value=None, lon_fill_value=None, fill_value=None,
-        do_single_sample=False):
+        do_single_sample=False, method='ewa'):
     """Remap data using the C or python version of ll2cr and the
     C version of fornav.
 
@@ -343,7 +497,7 @@ def remap_bands(sat, instrument, nav_set_uid, lon_fbf, lat_fbf,
     """
     # Used to determine verbosity
     log_level = logging.getLogger('').handlers[0].level or 0
-    
+
     # TODO, this is just a very rough check for now, in the long run handle this more gracefully
     assert (lat_fill_value == lon_fill_value)
     
@@ -355,10 +509,11 @@ def remap_bands(sat, instrument, nav_set_uid, lon_fbf, lat_fbf,
             lat_fill_value=lat_fill_value, lon_fill_value=lon_fill_value)
 
     # Run fornav
+    # FUTURE: 'method' should be used to determine entire method not just what type of fornav call to make
     fornav_output = run_fornav(sat, instrument, nav_set_uid, grid_jobs, ll2cr_output,
             num_procs=num_procs, verbose=log_level <= logging.DEBUG,
             fornav_d=fornav_d, fornav_D=fornav_D, fill_value=fill_value,
-            do_single_sample=do_single_sample)
+            do_single_sample=do_single_sample, method=method)
 
     return fornav_output
